requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6